refactor: openGraph 공통 메타데이터 상수로 분리
