+++ src/utils/creat-page-metadata.ts
@@ -6,11 +6,27 @@ interface CreatePageMetadataParams {
   path?: string;
   image?: string;
 }
+
+const DEFAULT_DESCRIPTION
+  = '공연이 만들어지고, 보여지고, 시작되는 곳. 버스킹의 모든 순간을 잇다, UNIBUSK';
+
+const DEFAULT_IMAGE = '/logos/logo-unibusk-stacked-vertical.png';
+
+export const SHARED_OPEN_GRAPH = {
+  siteName: 'UNIBUSK',
+  locale: 'ko_KR',
+  type: 'website' as const,
+};
+
+const SHARED_TWITTER = {
+  card: 'summary_large_image' as const,
+};
+
 export function createPageMetadata({
   title,
-  description = '버스킹의 모든 순간을 잇다, UNIBUSK',
+  description = DEFAULT_DESCRIPTION,
   path = '/',
-  image = '/logos/logo-unibusk-stacked-vertical.png',
+  image = DEFAULT_IMAGE,
 }: CreatePageMetadataParams): Metadata {
   return {
     title,
@@ -19,13 +35,14 @@ export function createPageMetadata({
       canonical: path,
     },
     openGraph: {
+      ...SHARED_OPEN_GRAPH,
       title,
       description,
       url: path,
       images: [image],
     },
     twitter: {
-      card: 'summary_large_image',
+      ...SHARED_TWITTER,
       title,
       description,
       images: [image],
