+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export { cn } from './cn';
+export { createPageMetadata, SHARED_OPEN_GRAPH } from './creat-page-metadata';
 export { ENV } from './environment';
 export { isSameOrderedIds } from './id-equals';
 export { routePaths } from './route-paths';
